Fix pylint errors in translation canary
These are new errors raised by the fact that we now testing even these
files by pylint.

The algorithm to search files for pylint have changed. Now all the files
without a pylint module and all pylint modules are tested.

Note:
We don't looking recursively into the pylint modules from now on. That
could be a problem but it's much faster.